Create a BaseService and ichange to BaseIngress to inherit from it
